いろいろ足りなかったし、考慮も漏れてた
