+++ Reversi/InitialPhase.swift
@@ -6,6 +6,21 @@ extension PhaseKind {
 
 public struct InitialPhase: Phase {
     public var kind: PhaseKind { .initial }
+
+    public func onExit(state: State, nextPhase: AnyPhase) -> State {
+        var state = state
+        
+        // ゲームの準備が整って始まったところでいろいろ初期化して外部に伝える
+        state.thinking = false
+        for disk in Disk.sides {
+            state.diskCount[disk] = state.board.countDisks(of: disk)
+        }
+        state.boardUpdateRequest = nil
+        state.passNotificationRequest = nil
+        state.resetConfirmationRequst = nil
+
+        return state
+    }
     
     public func reduce(state: State, action: Action) -> State {
         var reducedState = state
